+++ LdapTools/internal/scripts/variables.ps1
@@ -1,0 +1,2 @@
+﻿# Used in the module for capital-casing properties. Consumed by Get-LdapObject
+$script:culture = Get-Culture