+++ LdapTools/changelog.md
@@ -1,5 +1,13 @@
 ﻿# Changelog
+
+## 1.1.5 (2021-03-12)
+
++ New: Command Disable-LdapAccount - disables Active Directory accounts
++ New: Command Enable-LdapAccount - enables Active Directory accounts
++ New: Command Set-LdapAccountPassword - reset the password on the target AD account.
++ Upd: Get-LdapObject - added parameter alias "SearchBase" to SearchRoot parameter for AD module compatibility.
++ Upd: Get-LdapObject - improved output object property casing
+
 ## 1.0.0 (2020-09-17)
- - New: Some Stuff
- - Upd: Moar Stuff
- - Fix: Much Stuff
+
++ Initial Release